pyproject.toml
[project]
name = "fastbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.2.0",
    "fastapi[standard]>=0.115.4",
    "jinja2>=3.1.4",
    "passlib>=1.7.4",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.17",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.0",
]
